refactor(test): reuse Rego check
