+++ pkg/iac/scanners/terraform/setup_test.go
@@ -2,6 +2,7 @@ package terraform
 
 import (
 	"context"
+	"io/fs"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -13,6 +14,33 @@ import (
 	"github.com/aquasecurity/trivy/pkg/iac/scanners/terraform/parser"
 	"github.com/aquasecurity/trivy/pkg/iac/terraform"
 )
+
+var emptyBucketCheck = `# METADATA
+# schemas:
+# - input: schema.cloud
+# custom:
+#   avd_id: USER-TEST-0123
+#   short_code: non-empty-bucket
+#   provider: aws
+#   service: s3
+#   aliases:
+#   - my-alias
+#   input:
+#     selector:
+#     - type: cloud
+#       subtypes:
+#         - service: s3
+#           provider: aws
+package user.test123
+
+import rego.v1
+
+deny contains res if  {
+	some bucket in input.aws.s3.buckets
+	bucket.name.value == ""
+	res := result.new("The bucket name cannot be empty.", bucket.name)
+}
+`
 
 func createModulesFromSource(t *testing.T, source, ext string) terraform.Modules {
 	fs := testutil.CreateFS(t, map[string]string{
@@ -30,14 +58,26 @@ func createModulesFromSource(t *testing.T, source, ext string) terraform.Modules
 	return modules
 }
 
+func scanFS(fsys fs.FS, target string, opts ...options.ScannerOption) (scan.Results, error) {
+	s := New(append(
+		[]options.ScannerOption{
+			rego.WithEmbeddedLibraries(true),
+			options.ScannerWithRegoOnly(true),
+			ScannerWithAllDirectories(true),
+		},
+		opts...,
+	)...,
+	)
+
+	return s.ScanFS(context.TODO(), fsys, target)
+}
+
 func scanHCL(t *testing.T, source string, opts ...options.ScannerOption) scan.Results {
 
-	fs := testutil.CreateFS(t, map[string]string{
+	fsys := testutil.CreateFS(t, map[string]string{
 		"main.tf": source,
 	})
-
-	localScanner := New(append(opts, rego.WithEmbeddedPolicies(false))...)
-	results, err := localScanner.ScanFS(context.TODO(), fs, ".")
+	results, err := scanFS(fsys, ".", opts...)
 	require.NoError(t, err)
 	return results
 }
